refactoring CategoryData type
moved CategoryData type into a separated file to be able to use it every where
