+++ lib/features/home/logic/cubit/home_state.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:waie/core/networking/api_error_handler.dart';
-import 'package:waie/features/home/data/model/category_response.dart';
+import 'package:waie/core/shared_models/category_data_model/category_data.dart';
 
 part 'home_state.freezed.dart';
